Move all RCF examples to apps/docs
