+++ apps/docs/docs/react-composable-form/concepts/Layout.mdx
@@ -3,8 +3,8 @@ sidebar_position: 2
 hide_table_of_contents: true
 ---
 
-import * as component from "./Layout.components";
-import * as code from "!!slice-loader!./Layout.components";
+import * as component from "./Layout";
+import * as code from "!!slice-loader!./Layout";
 import { Split } from "../../../src/components/Split";
 import CodeBlock from "@theme/CodeBlock";
