+++ apps/docs/docs/react-composable-form/concepts/Remote.mdx
@@ -1,0 +1,5 @@
+import RemoteExample from "./Remote";
+
+# Remote
+
+<RemoteExample />
