+++ apps/docs/docs/react-composable-form/examples/PropertyExplorer.mdx
@@ -1,0 +1,1 @@
+# Property Explorer
