+++ apps/docs/docs/react-composable-form/examples/Search.mdx
@@ -1,1 +1,5 @@
+import Search from "./Search";
+
 # Search
+
+<Search />
