+++ apps/docs/docs/react-composable-form/index.mdx
@@ -2,9 +2,9 @@
 sidebar_position: 1
 ---
 
-import { UserForm } from "./index.components";
+import { UserForm } from "./index";
 import { Split } from "../../src/components/Split";
-import codeElements from "!!shiki-twoslash-loader!./index.components";
+import codeElements from "!!shiki-twoslash-loader!./index";
 
 # Overview
