+++ apps/docs/package.json
@@ -14,7 +14,8 @@
     "react": "18.2.0",
     "react-composable-form": "workspace:*",
     "react-dom": "18.2.0",
-    "react-imperative-hook": "workspace:*"
+    "react-imperative-hook": "workspace:*",
+    "react-query": "^3.39.3"
   },
   "devDependencies": {
     "@docusaurus/core": "3.0.0-alpha.0",
