--- apps/rcf-examples/src/examples/CheckoutExample.tsx
@@ -1,3 +1,0 @@
-export function CheckoutExample() {
-  return null;
-}
