--- apps/rcf-examples/src/examples/DataFormatExample.tsx
@@ -1,3 +1,0 @@
-export function DataFormatExample() {
-  return null;
-}
