--- apps/rcf-examples/src/examples/ListExample.tsx
@@ -1,3 +1,0 @@
-export function ListExample() {
-  return null;
-}
