--- apps/rcf-examples/src/examples/NestedExample.tsx
@@ -1,3 +1,0 @@
-export function NestedExample() {
-  return null;
-}
