--- apps/rcf-examples/src/examples/PropertyExplorerExample.tsx
@@ -1,3 +1,0 @@
-export function PropertyExplorerExample() {
-  return null;
-}
