--- apps/rcf-examples/src/examples/StepsExample.tsx
@@ -1,3 +1,0 @@
-export function StepsExample() {
-  return null;
-}
