+++ apps/rcf-examples/src/main.tsx
@@ -1,9 +1,8 @@
 import { createRoot } from "react-dom/client";
 import { StrictMode } from "react";
-import { App } from "./App";
 
 createRoot(document.getElementById("root") as HTMLElement).render(
   <StrictMode>
-    <App />
+    <div>Hello World</div>
   </StrictMode>,
 );
